[CRT-434] Merge TaskInstance and ProgressList
