+++ frontend/src/hooks/useStudentProgress.ts
@@ -1,6 +1,8 @@
 import { useMemo } from "react";
 import { ClassStudent } from "@/api/collimator/models/classes/class-student";
 import { ExistingClassExtended } from "@/api/collimator/models/classes/existing-class-extended";
+import { useClass } from "@/api/collimator/hooks/classes/useClass";
+import { useClassSession } from "@/api/collimator/hooks/sessions/useClassSession";
 
 type AnonymousStudent = {
   isAnonymous: true;
@@ -35,3 +37,33 @@ export const useStudentProgress = (
         } satisfies AnonymousStudent),
     );
   }, [klass, activeStudentIds]);
+
+export const useSessionStudents = (
+  classId: number,
+  sessionId: number,
+  activeStudentIds: number[],
+) => {
+  const {
+    data: klass,
+    error: klassError,
+    isLoading: isLoadingKlass,
+  } = useClass(classId);
+
+  const {
+    data: session,
+    error: sessionError,
+    isLoading: isLoadingSession,
+  } = useClassSession(classId, sessionId);
+
+  const students = useStudentProgress(klass, activeStudentIds);
+
+  return {
+    klass,
+    session,
+    students,
+    klassError,
+    sessionError,
+    isLoadingKlass,
+    isLoadingSession,
+  };
+};
